+++ Library/tests/Contracts/DecimalTests.cs
@@ -2,13 +2,11 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using NUnit.Framework;
-using Phantasma.Core.Cryptography;
-using Phantasma.Core.Domain;
+using PhantasmaPhoenix.Cryptography;
+using PhantasmaPhoenix.Protocol;
 using Phantasma.Core.Domain.Contract;
-using Phantasma.Core.Domain.Execution.Enums;
-using Phantasma.Core.Domain.VM;
-using Phantasma.Core.Numerics;
-using Phantasma.Core.Utils;
+using PhantasmaPhoenix.VM;
+using PhantasmaPhoenix.Core;
 using Phantasma.Tomb;
 using Phantasma.Tomb.Compilers;
 
@@ -67,7 +65,7 @@ public class DecimalTests
 
         var obj = vm.Stack.Pop();
         var newVal = obj.AsNumber();
-        var expectedVal = UnitConversion.ToBigInteger(val, decimals);
+        var expectedVal = UnitConversion.ToBigInteger(val, (uint)decimals);
 
         Assert.IsTrue(newVal == expectedVal);
